Add has_one_base64_attached method
