+++ lib/active_storage_base64.rb
@@ -1,0 +1,3 @@
+require 'active_storage_support/support_for_base64'
+require 'active_storage_support/base64_one'
+require 'active_storage_support/base64_attach'
